Merged PR 409: Allows to active / deactivate the content which is in Transformed or Broadcasted state
Allows the super admin to change the content status to active or inactive. It is allowed only if the content is in Transformed or Broadcasted state. In case, in some scenario, admin want to stop the online streaming of some content, it can be done by changing the flag to false.

Related work items: #3152