Merged PR 455: Fix authorizations in Retailer Controller
## Issue

Retailer is not able to access /me API

Reason: Controller is marked with SuperAdmin role, whereas the me API is marked with Retailer role. the API marking is not an override, but is supposed to be a subset of the controller marking. Due to this, a user with only Retailer role cannot access the /me API.

## Fix

Updated the controller marking as well as each API marking to reflect the correct role authorizations.

Related work items: #3361