Merged PR 528: Device to Cloud Communication
Taken a change in approach for Device to Cloud communication as the existing approach of invoking the Cloud API directly from the device was bringing Authentication complexities with Kaizala and was requiring different service accounts from Kaizala. Kaizala is not designed to scale with many service account. Hence, the change in approach and in the new approach we send the Operations to execute as a part of Telemetry data to IOT central and then publish the same to service bus.

Once the data is received in service bus, we process it by looking at the Command Name and Command Data. This approach has the dependency on IOT Central and needs to be ensured that we have right export set at IOT central level.

Related work items: #3522