Merged PR 414: Move Cloud API and Listener containers to Mariner
## Issue

From compliance review, AKS + Ubuntu was flagged as a risk. As a mitigation, AKS + Mariner and Mariner containers was recommended.

## Changes

This change is to migrate the container images to CBL-Mariner (1st party Linux distro by MSFT). This will be required to migrate to AKS+Mariner host deployment, and will later be a prerequisite for COSMIC on-boarding.

Tags reference:
https://mcr.microsoft.com/v2/dotnet/sdk/tags/list
https://mcr.microsoft.com/v2/dotnet/runtime/tags/list
https://mcr.microsoft.com/v2/dotnet/aspnet/tags/list

## Validations

The User API container was deployed on dev environment and no issues found with user and auth workflows.

Related work items: #3546