Merged PR 653: Rollback the change. No database update reqd. if time stamp is same.
Rollback the change. no database update reqd. if time stamp is same. Ignore the asset map records with the same time stamp. IF we keep = every 24 hours we will have an update to COSMOS DB even if record has not changed and it will result in unnecessary high RU

Related work items: #3658