Merged PR 670: Added a method to return the content availability count on the given device ids
Added a method to return the content availability count on the given device ids.
This is an approximate count.
In case, a broadcast has been cancelled, it will take some time to get reflected here.
Currently there is a limitation from SES if device is Powered OFF for 24 hours after cancellation, it will never get notified here

Related work items: #3779