Merged PR 794: Fix build issue in Web Portal dockerfile
## Issue

A recent update to nodejs package was published in Mariner to get rid of the cipher override. This is now the default nodejs installation.

However, the corresponding CA certificates package is missing as Mariner core image does not install CA certificates.

This is causing `npm install` to fail during build

## Fix

Add CA certificate package installation in the Dockerfile

## Validation

- Build passes
- Able to run the container locally and Portal webpage shows up as usual

Related work items: #4032