Merged PR 1352: Support of Order Completed Event Across Content Provider
Currently incentive system allows setting up the incentives for the Order Complete Event along with the content provider only. For example: you can set up an Incentive Plan to provide a commission of 10 % on the completion of Eros Order but the system does not allow setting up 10 % of commission on order complete of ANY content provider.

The current PR allows setting up Incentive without choosing the content provider. Also, on order complete now 4 Incentive events are generated instead of 2. 2 for retailer 2 for consumer (with and without incentive sub type).

Related work items: #5419